refactor stdin code
