+++ src/vs/code/node/cli.ts
@@ -14,10 +14,10 @@ import product from 'vs/platform/product/common/product';
 import * as paths from 'vs/base/common/path';
 import { whenDeleted, writeFileSync } from 'vs/base/node/pfs';
 import { findFreePort, randomPort } from 'vs/base/node/ports';
-import { resolveTerminalEncoding } from 'vs/base/node/encoding';
 import { isWindows, isLinux } from 'vs/base/common/platform';
 import { ProfilingSession, Target } from 'v8-inspect-profiler';
 import { isString } from 'vs/base/common/types';
+import { hasStdinWithoutTty, stdinDataListener, getStdinFilePath, readFromStdin } from 'vs/platform/environment/node/stdin';
 
 function shouldSpawnCliProcess(argv: ParsedArgs): boolean {
 	return !!argv['install-source']
@@ -142,91 +142,55 @@ export async function main(argv: string[]): Promise<any> {
 			});
 		}
 
-		let stdinWithoutTty: boolean = false;
-		try {
-			stdinWithoutTty = !process.stdin.isTTY; // Via https://twitter.com/MylesBorins/status/782009479382626304
-		} catch (error) {
-			// Windows workaround for https://github.com/nodejs/node/issues/11656
-		}
-
-		const readFromStdin = args._.some(a => a === '-');
-		if (readFromStdin) {
+		const hasReadStdinArg = args._.some(a => a === '-');
+		if (hasReadStdinArg) {
 			// remove the "-" argument when we read from stdin
 			args._ = args._.filter(a => a !== '-');
 			argv = argv.filter(a => a !== '-');
 		}
 
-		let stdinFilePath: string;
-		if (stdinWithoutTty) {
+		let stdinFilePath: string | undefined;
+		if (hasStdinWithoutTty()) {
 
 			// Read from stdin: we require a single "-" argument to be passed in order to start reading from
 			// stdin. We do this because there is no reliable way to find out if data is piped to stdin. Just
 			// checking for stdin being connected to a TTY is not enough (https://github.com/Microsoft/vscode/issues/40351)
-			if (args._.length === 0 && readFromStdin) {
 
-				// prepare temp file to read stdin to
-				stdinFilePath = paths.join(os.tmpdir(), `code-stdin-${Math.random().toString(36).replace(/[^a-z]+/g, '').substr(0, 3)}.txt`);
+			if (args._.length === 0) {
+				if (hasReadStdinArg) {
+					stdinFilePath = getStdinFilePath();
 
-				// open tmp file for writing
-				let stdinFileError: Error | undefined;
-				let stdinFileStream: fs.WriteStream;
-				try {
-					stdinFileStream = fs.createWriteStream(stdinFilePath);
-				} catch (error) {
-					stdinFileError = error;
-				}
+					// returns a file path where stdin input is written into (write in progress).
+					try {
+						readFromStdin(stdinFilePath, !!verbose); // throws error if file can not be written
 
-				if (!stdinFileError) {
+						// Make sure to open tmp file
+						addArg(argv, stdinFilePath);
 
-					// Pipe into tmp file using terminals encoding
-					resolveTerminalEncoding(verbose).then(async encoding => {
-						const iconv = await import('iconv-lite');
-						const converterStream = iconv.decodeStream(encoding);
-						process.stdin.pipe(converterStream).pipe(stdinFileStream);
-					});
+						// Enable --wait to get all data and ignore adding this to history
+						addArg(argv, '--wait');
+						addArg(argv, '--skip-add-to-recently-opened');
+						args.wait = true;
 
-					// Make sure to open tmp file
-					addArg(argv, stdinFilePath);
-
-					// Enable --wait to get all data and ignore adding this to history
-					addArg(argv, '--wait');
-					addArg(argv, '--skip-add-to-recently-opened');
-					args.wait = true;
-				}
-
-				if (verbose) {
-					if (stdinFileError) {
-						console.error(`Failed to create file to read via stdin: ${stdinFileError.toString()}`);
-					} else {
 						console.log(`Reading from stdin via: ${stdinFilePath}`);
+					} catch (e) {
+						console.log(`Failed to create file to read via stdin: ${e.toString()}`);
+						stdinFilePath = undefined;
 					}
-				}
-			}
+				} else {
 
-			// If the user pipes data via stdin but forgot to add the "-" argument, help by printing a message
-			// if we detect that data flows into via stdin after a certain timeout.
-			else if (args._.length === 0) {
-				processCallbacks.push(child => new Promise(c => {
-					const dataListener = () => {
-						if (isWindows) {
-							console.log(`Run with '${product.applicationName} -' to read output from another program (e.g. 'echo Hello World | ${product.applicationName} -').`);
-						} else {
-							console.log(`Run with '${product.applicationName} -' to read from stdin (e.g. 'ps aux | grep code | ${product.applicationName} -').`);
+					// If the user pipes data via stdin but forgot to add the "-" argument, help by printing a message
+					// if we detect that data flows into via stdin after a certain timeout.
+					processCallbacks.push(_ => stdinDataListener(1000).then(dataReceived => {
+						if (dataReceived) {
+							if (isWindows) {
+								console.log(`Run with '${product.applicationName} -' to read output from another program (e.g. 'echo Hello World | ${product.applicationName} -').`);
+							} else {
+								console.log(`Run with '${product.applicationName} -' to read from stdin (e.g. 'ps aux | grep code | ${product.applicationName} -').`);
+							}
 						}
-
-						c(undefined);
-					};
-
-					// wait for 1s maximum...
-					setTimeout(() => {
-						process.stdin.removeListener('data', dataListener);
-
-						c(undefined);
-					}, 1000);
-
-					// ...but finish early if we detect data
-					process.stdin.once('data', dataListener);
-				}));
+					}));
+				}
 			}
 		}
